requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
sqlalchemy==2.0.35
alembic==1.13.2
openai==1.12.0
Pillow==10.3.0
python-dotenv==1.0.1
aiohttp==3.9.5
pydantic==2.6.1
pydantic-settings==2.1.0

